move Show answer button from App to Question section so it resets on question change
